Grid functions added and some files deleted
Add the capability to generate a grid of a specified area across a set of points, count how many points are in each cell and check that there aren't more points than specified, then split cells if needed to get them below the threshold number of points. Finally, for each grid cell, the points that overlap with it are exported.

Next, added the ability to extract imagery band information points and from the file containing each set of points for each grid cell.