Resampling and masking modification
Are now capable of resampling imagery up or down depending on the need and use case. Images can now be masked to contain data within polygon-defined areas (e.g. where the shrubs are). To save space prior to merging images together, it is recommend to run the model on each of the masked images.